+++ src/IconButton/IconButton.tsx
@@ -33,29 +33,31 @@ export enum IconButtonTestIds {
   ICONBUTTON_END_ICON = 'iconButtonEndIcon',
 }
 
-const StyledMainContainer = styled('div')((theme) => {
+const StyledMainContainer = styled('div')<{ darkMode?: boolean }>((props) => {
+  const { theme, darkMode } = props;
   return {
     flexDirection: 'column',
     justifyContent: 'center',
     display: 'inline-flex',
     alignItems: 'center',
     '&.selected': {
-      color: theme.theme.palette.primary.main,
+      color: darkMode ? theme.palette.action.selectedInverse : theme.palette.action.selected,
       '.MuiSvgIcon-root, .MuiTypography-root': {
-        color: theme.theme.palette.primary.main,
+        color: darkMode ? theme.palette.action.selectedInverse : theme.palette.action.selected,
       },
     },
     '&.disabled': {
-      color: theme.theme.palette.action.disabled,
+      color: darkMode ? theme.palette.action.disabledInverse : theme.palette.action.disabled,
       '.MuiSvgIcon-root, .MuiTypography-root': {
-        color: theme.theme.palette.action.disabled,
+        color: darkMode ? theme.palette.action.disabledInverse : theme.palette.action.disabled,
       },
       pointerEvents: 'none',
     },
   };
 });
 
-const StyledSubContainer = styled('div')((theme) => {
+const StyledSubContainer = styled('div')<{ darkMode?: boolean }>((props) => {
+  const { theme, darkMode } = props;
   return {
     flexDirection: 'row',
     justifyContent: 'center',
@@ -64,35 +66,35 @@ const StyledSubContainer = styled('div')((theme) => {
     width: 'fit-content',
     '&:hover': {
       borderRadius: '2px',
-      backgroundColor: theme.theme.palette.action.hover,
+      backgroundColor: darkMode ? theme.palette.action.hoverInverse : theme.palette.action.hover,
     },
     '&.force-to-focusHover': {
       borderRadius: '2px',
-      backgroundColor: theme.theme.palette.action.hover,
+      backgroundColor: darkMode ? theme.palette.action.hoverInverse : theme.palette.action.hover,
     },
     '&.selected': {
-      outline: `1px solid ${theme.theme.palette.action.focus}`,
+      outline: `1px solid ${darkMode ? theme.palette.action.focusInverse : theme.palette.action.focus}`,
       borderRadius: '1px',
-      backgroundColor: theme.theme.palette.action.selectedOpacityModified,
+      backgroundColor: theme.palette.action.selectedOpacityModified,
       '&:hover': {
-        backgroundColor: theme.theme.palette.action.selectedOpacityHover,
-        color: theme.theme.palette.primary.dark,
-        outline: `1px solid ${theme.theme.palette.primary.dark}`,
+        backgroundColor: theme.palette.action.selectedOpacityModified,
+        color: darkMode ? theme.palette.primary.darkInverse : theme.palette.primary.dark,
+        outline: `1px solid ${darkMode ? theme.palette.primary.darkInverse : theme.palette.primary.dark}`,
         '.MuiSvgIcon-root, + .MuiTypography-root': {
-          color: theme.theme.palette.primary.dark,
+          color: darkMode ? theme.palette.primary.darkInverse : theme.palette.primary.dark,
         },
       },
       '&.force-to-focusHover': {
-        backgroundColor: theme.theme.palette.action.selectedOpacityHover,
-        color: theme.theme.palette.primary.dark,
-        outline: `1px solid ${theme.theme.palette.primary.dark}`,
+        backgroundColor: theme.palette.action.selectedOpacityModified,
+        color: darkMode ? theme.palette.primary.darkInverse : theme.palette.primary.dark,
+        outline: `1px solid ${darkMode ? theme.palette.primary.darkInverse : theme.palette.primary.dark}`,
         '.MuiSvgIcon-root, + .MuiTypography-root': {
-          color: theme.theme.palette.primary.dark,
+          color: darkMode ? theme.palette.primary.darkInverse : theme.palette.primary.dark,
         },
       },
       '&.disabled': {
-        backgroundColor: theme.theme.palette.action.disabledOpacityModified,
-        outline: `1px solid ${theme.theme.palette.border.secondary}`,
+        backgroundColor: theme.palette.action.disabledOpacityModified,
+        outline: `1px solid ${darkMode ? theme.palette.border.inverseSecondary : theme.palette.border.secondary}`,
         borderRadius: '1px',
       },
     },
@@ -105,22 +107,22 @@ export const getMuiIconButtonThemeOverrides = (): Components<Omit<Theme, 'compon
       styleOverrides: {
         root: ({ ownerState, theme }) => {
           return ({
-            color: theme.palette.action.active,
+            color: ownerState.darkMode ? theme.palette.action.inverse : theme.palette.action.active,
             backgroundColor: 'transparent',
             borderRadius: '2px',
             padding: 0,
             '&.Mui-focusVisible, &:focus, &.force-to-focus, &.force-to-focusHover': {
-              outline: `1px solid ${theme.palette.action.focus}`,
+              outline: `1px solid ${ownerState.darkMode ? theme.palette.action.focusInverse : theme.palette.action.focus}`,
               borderRadius: '3px',
               outlineOffset: '-2px',
             },
             '&:hover': {
               borderRadius: '2px',
-              backgroundColor: theme.palette.action.hover,
+              backgroundColor: ownerState.darkMode ? theme.palette.action.hoverInverse : theme.palette.action.hover,
             },
             '&.force-to-focusHover': {
               borderRadius: '2px',
-              backgroundColor: theme.palette.action.hover,
+              backgroundColor: ownerState.darkMode ? theme.palette.action.hoverInverse : theme.palette.action.hover,
             },
             '.MuiSvgIcon-root:not(.endIcon)': {
               margin: '0',
@@ -163,15 +165,18 @@ export type IconButtonProps = MuiIconButtonProps & {
   selected?: boolean,
   label?: string,
   showendicon?: boolean | 0 | 1,
+  darkMode?: boolean,
 }
 
 const IconButton = React.forwardRef(({ showendicon, ...props }: IconButtonProps, forwardRef) => {
   return (
     <StyledMainContainer
       className={`IconButtonMainContainer ${props.selected ? 'selected' : ''} ${props.disabled ? 'disabled' : ''} ${props.className}`}
+      darkMode={props.darkMode}
     >
       <StyledSubContainer
         className={`${props.selected ? 'selected' : ''} ${props.disabled ? 'disabled' : ''} ${props.className}`}
+        darkMode={props.darkMode}
       >
         <MuiIconButton
           {...props}
@@ -187,7 +192,7 @@ const IconButton = React.forwardRef(({ showendicon, ...props }: IconButtonProps,
               data-testid={IconButtonTestIds.ICONBUTTON_END_ICON}
               sx={(theme) => {
                 return {
-                  color: theme.palette.action.active,
+                  color: props.darkMode ? theme.palette.action.inverse : theme.palette.action.active,
                   width: '12px',
                   height: '12px',
                   margin: '0',
@@ -221,7 +226,7 @@ const IconButton = React.forwardRef(({ showendicon, ...props }: IconButtonProps,
         textAlign="center"
         sx={(theme) => {
           return {
-            color: theme.palette.action.active,
+            color: props.darkMode ? theme.palette.action.inverse : theme.palette.action.active,
             paddingLeft: '4px',
             paddingRight: '4px',
             marginTop: '2px',
@@ -243,6 +248,7 @@ IconButton.defaultProps = {
   showendicon: false,
   label: undefined,
   disabled: false,
+  darkMode: false,
   disableFocusRipple: true,
   edge: false,
   centerRipple: true,