+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## Unreleased
 
 ### Added
+- Added an inverse color variant to the `IconButton` component for dark mode
+- Added the `darkInverse`, `focusInverse`, and `selectedInverse` colors to the palette
 
 ### Fixed
+- Corrected the focus styling of `preview icon` in the `Tile` component and `zoom buttons` in the `Preview` component to meet the required contrast ratio
 
 ### Changed
 - Added keyboard accessibility to the header in `DataGrid` component